refactor(time): isolate browser-safe duration parts
Move the existing duration quantities unchanged and keep presentation adapters separate. Workboard's supported SDK seam remains a maintainer decision at the current public-entrypoint cap.

Proof: focused duration, UI, and Workboard suites passed 433 tests. Independent review found no actionable P0-P2 findings.
